Deprecate build_driz_weight -> stcal; move load_custom_wcs -> utils
