+++ jwst/resample/resample_step.py
@@ -1,10 +1,8 @@
 import logging
-from copy import deepcopy
-
-import asdf
 
 from jwst.datamodels import ModelLibrary, ImageModel  # type: ignore[attr-defined]
 from jwst.lib.pipe_utils import match_nans_and_flags
+from jwst.resample.resample_utils import load_custom_wcs
 
 from . import resample
 from ..stpipe import Step
@@ -199,60 +197,3 @@ class ResampleStep(Step):
             self.log.debug('   {}={}'.format(k, v))
 
         return kwargs
-
-
-def load_custom_wcs(asdf_wcs_file, output_shape=None):
-    """
-    Load a custom output WCS from an ASDF file.
-
-    Parameters
-    ----------
-    asdf_wcs_file : str
-        Path to an ASDF file containing a GWCS structure.
-    output_shape : tuple of int, optional
-        Array shape for the output data.  If not provided,
-        the custom WCS must specify one of: pixel_shape,
-        array_shape, or bounding_box.
-
-    Returns
-    -------
-    wcs : WCS
-        The output WCS to resample into.
-    """
-    if not asdf_wcs_file:
-        return None
-
-    with asdf.open(asdf_wcs_file) as af:
-        wcs = deepcopy(af.tree["wcs"])
-        pixel_area = af.tree.get("pixel_area", None)
-        pixel_shape = af.tree.get("pixel_shape", None)
-        array_shape = af.tree.get("array_shape", None)
-
-    if not hasattr(wcs, "pixel_area") or wcs.pixel_area is None:
-        wcs.pixel_area = pixel_area
-    if not hasattr(wcs, "pixel_shape") or wcs.pixel_shape is None:
-        wcs.pixel_shape = pixel_shape
-    if not hasattr(wcs, "array_shape") or wcs.array_shape is None:
-        wcs.array_shape = array_shape
-
-    if output_shape is not None:
-        wcs.array_shape = output_shape[::-1]
-        wcs.pixel_shape = output_shape
-    elif wcs.pixel_shape is not None:
-        wcs.array_shape = wcs.pixel_shape[::-1]
-    elif wcs.array_shape is not None:
-        wcs.pixel_shape = wcs.array_shape[::-1]
-    elif wcs.bounding_box is not None:
-        wcs.array_shape = tuple(
-            int(axs[1] + 0.5)
-            for axs in wcs.bounding_box.bounding_box(order="C")
-        )
-        wcs.pixel_shape = wcs.array_shape[::-1]
-    else:
-        raise ValueError(
-            "Step argument 'output_shape' is required when custom WCS "
-            "does not have 'array_shape', 'pixel_shape', or "
-            "'bounding_box' attributes set."
-        )
-
-    return wcs
